Location Class needs to use Double
having problems with Floats and retrieving stuff from the database,
changed to Double and it seems to be working now

+++ app/models/Location.scala
@@ -21,7 +21,7 @@ import globals.ActiveDatabase
  * @param latitude The latitude of the location. (-90 <= _ <= 90)
  * @param longitude The longitude of the location (-180 < _ <= 180)
  */
-case class Location(id: Int, name: String, latitude: Float, longitude: Float, address: String,
+case class Location(id: Int, name: String, latitude: Double, longitude: Double, address: String,
 	postalCode: String, city: String, regionalHealthAuthority: String, inspections: Seq[Inspection]) {
   /** Returns true if the location has at least one inspection. */
   def hasInspections: Boolean = inspections.size != 0
@@ -121,7 +121,7 @@ object Location {
   private def locationRowToLocation(row: Row)(implicit connection: java.sql.Connection): Try[Location] = {
     Inspection.getInspections(row[Int]("id")) match {
       case Success(inspections) =>
-        Success(Location(row[Int]("id"), row[String]("name"), row[Float]("latitude"), row[Float]("longitude"), 
+        Success(Location(row[Int]("id"), row[String]("name"), row[Double]("latitude"), row[Double]("longitude"), 
             row[String]("address"), row[String]("postcode"), row[String]("city"), row[String]("rha"),  inspections))
       case Failure(ex) => 
         Failure(ex)
